backend/requirements.txt
fastapi
uvicorn
claude-agent-sdk
python-dotenv
sqlalchemy
pydantic-settings
loguru
